+++ src/messageExtension/templates/src/app/{messageExtensionName}/{messageExtensionClassName}.ts
@@ -1,77 +1,54 @@
 import * as debug from "debug";
-import { IMessageExtension, PreventIframe } from "express-msteams-host";
-import { TurnContext, CardFactory, InvokeResponse } from "botbuilder";
-import { MessagingExtensionQuery, InvokeResponseTypeOf, InvokeResponseTyped, MessagingExtensionResponse } from "botbuilder-teams";
+import { PreventIframe } from "express-msteams-host";
+import { TurnContext, CardFactory } from "botbuilder";
+import { MessagingExtensionQuery, MessagingExtensionResult } from "botbuilder-teams";
+import { IMessagingExtensionMiddlewareProcessor } from "botbuilder-teams-messagingextensions";
 
 
 // Initialize debug logging module
 const log = debug("msteams");
 
 @PreventIframe("/<%=messageExtensionName%>/config.html")
-export default class <%= messageExtensionClassName%> implements IMessageExtension {
+export default class <%= messageExtensionClassName%> implements IMessagingExtensionMiddlewareProcessor {
 
-    public async onQuery(context: TurnContext, query: MessagingExtensionQuery): Promise<InvokeResponseTyped<MessagingExtensionResponse>> {
-        type R = InvokeResponseTypeOf<"onMessagingExtensionQuery">;
+    public async onQuery(context: TurnContext, query: MessagingExtensionQuery): Promise<MessagingExtensionResult> {
         const card = CardFactory.heroCard("Test", "Test", ["<%=host%>/assets/icon.png"]);
 
         if (query.parameters && query.parameters[0] && query.parameters[0].name === "initialRun") {
             // initial run
 
             return Promise.resolve({
-                status: 200,
-                body: {
-                    composeExtension: {
-                        type: "result",
-                        attachmentLayout: "list",
-                        attachments: [
-                            card
-                        ]
-                    }
-                }
-            } as R);
+                type: "result",
+                attachmentLayout: "grid",
+                attachments: [
+                    card
+                ]
+            } as MessagingExtensionResult);
         } else {
             // the rest
             return Promise.resolve({
-                status: 200,
-                body: {
-                    composeExtension: {
-                        type: "result",
-                        attachmentLayout: "list",
-                        attachments: [
-                            card
-                        ]
-                    }
-                }
-            } as R);
+                type: "result",
+                attachmentLayout: "list",
+                attachments: [
+                    card
+                ]
+            } as MessagingExtensionResult);
         }
     }
 
     // this is used when canUpdateConfiguration is set to true
-    public async onQuerySettingsUrl(context: TurnContext): Promise<InvokeResponseTyped<{ composeExtension: { type: string, suggestedActions: { actions: Array<{ type: string, title: string, value: string }> } } }>> {
+    public async onQuerySettingsUrl(context: TurnContext): Promise<{ title: string, value: string }> {
         return Promise.resolve({
-            status: 200,
-            body: {
-                composeExtension: {
-                    type: "config",
-                    suggestedActions: {
-                        actions: [
-                            {
-                                type: "openApp",
-                                title: "<%=messageExtensionTitle%> Configuration",
-                                value: "<%= host %>/<%= messageExtensionName %>/config.html"
-                            }
-                        ]
-                    }
-                }
-            }
+            title: "<%=messageExtensionTitle%> Configuration",
+            value: "<%= host %>/<%= messageExtensionName %>/config.html"
         });
     }
 
-    public async onSettingsUpdate(context: TurnContext): Promise<InvokeResponse> {
+    public async onSettings(context: TurnContext): Promise<void> {
         // take care of the setting returned from the dialog, with the value stored in state
         const setting = context.activity.value.state;
         log(`New setting: ${setting}`);
-        return Promise.resolve({ status: 200 });
+        return Promise.resolve();
     }
 
 }